+++ .chloggen/codeboten_obsreport-scraper.yaml
@@ -1,0 +1,32 @@
+# Use this changelog template to create an entry for release notes.
+
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: deprecation
+
+# The name of the component, or a single word describing the area of concern, (e.g. otlpreceiver)
+component: obsreport
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: "Deprecating obsreport scraper and receiver API" 
+
+# One or more tracking issues or pull requests related to the change
+issues: [8492]
+
+# (Optional) One or more lines of additional information to render under the primary note.
+# These lines will be padded with 2 spaces and then inserted directly into the document.
+# Use pipe (|) for multiline entries.
+subtext: |
+  These deprecated methods/structs have been moved to receiverhelper and scraperhelper:
+  - `obsreport.Receiver` -> `receiverhelper.ObsReport`
+  - `obsreport.ReceiverSettings` -> `receiverhelper.ObsReportSettings`
+  - `obsreport.NewReceiver` -> `receiverhelper.NewObsReport`
+  - `obsreport.Scraper` -> `scraperhelper.ObsReport`
+  - `obsreport.ScraperSettings` -> `scraperhelper.ObsReportSettings`
+  - `obsreport.NewScraper` -> `scraperhelper.NewObsReport`
+
+# Optional: The change log or logs in which this entry should be included.
+# e.g. '[user]' or '[user, api]'
+# Include 'user' if the change is relevant to end users.
+# Include 'api' if there is a change to a library API.
+# Default: '[user]'
+change_logs: [api]
