--- obsreport/obsreport.go
@@ -1,21 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package obsreport // import "go.opentelemetry.io/collector/obsreport"
-
-import (
-	"go.opentelemetry.io/otel/codes"
-	"go.opentelemetry.io/otel/trace"
-)
-
-const (
-	scopeName = "go.opentelemetry.io/collector/obsreport"
-
-	nameSep = "/"
-)
-
-func recordError(span trace.Span, err error) {
-	if err != nil {
-		span.SetStatus(codes.Error, err.Error())
-	}
-}
