+++ receiver/otlpreceiver/internal/metrics/otlp.go
@@ -7,8 +7,8 @@ import (
 	"context"
 
 	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/obsreport"
 	"go.opentelemetry.io/collector/pdata/pmetric/pmetricotlp"
+	"go.opentelemetry.io/collector/receiver/receiverhelper"
 )
 
 const dataFormatProtobuf = "protobuf"
@@ -17,14 +17,14 @@ const dataFormatProtobuf = "protobuf"
 type Receiver struct {
 	pmetricotlp.UnimplementedGRPCServer
 	nextConsumer consumer.Metrics
-	obsrecv      *obsreport.Receiver
+	obsreport    *receiverhelper.ObsReport
 }
 
 // New creates a new Receiver reference.
-func New(nextConsumer consumer.Metrics, obsrecv *obsreport.Receiver) *Receiver {
+func New(nextConsumer consumer.Metrics, obsreport *receiverhelper.ObsReport) *Receiver {
 	return &Receiver{
 		nextConsumer: nextConsumer,
-		obsrecv:      obsrecv,
+		obsreport:    obsreport,
 	}
 }
 
@@ -36,9 +36,9 @@ func (r *Receiver) Export(ctx context.Context, req pmetricotlp.ExportRequest) (p
 		return pmetricotlp.NewExportResponse(), nil
 	}
 
-	ctx = r.obsrecv.StartMetricsOp(ctx)
+	ctx = r.obsreport.StartMetricsOp(ctx)
 	err := r.nextConsumer.ConsumeMetrics(ctx, md)
-	r.obsrecv.EndMetricsOp(ctx, dataFormatProtobuf, dataPointCount, err)
+	r.obsreport.EndMetricsOp(ctx, dataFormatProtobuf, dataPointCount, err)
 
 	return pmetricotlp.NewExportResponse(), err
 }
