+++ receiver/receiverhelper/obsreport_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package obsreport
+package receiverhelper
 
 import (
 	"context"
@@ -16,7 +16,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/internal/obsreportconfig/obsmetrics"
 	"go.opentelemetry.io/collector/obsreport/obsreporttest"
-	"go.opentelemetry.io/collector/receiver/scrapererror"
 )
 
 const (
@@ -26,33 +25,13 @@ const (
 
 var (
 	receiverID = component.NewID("fakeReceiver")
-	scraperID  = component.NewID("fakeScraper")
 
-	errFake        = errors.New("errFake")
-	partialErrFake = scrapererror.NewPartialScrapeError(errFake, 1)
+	errFake = errors.New("errFake")
 )
 
 type testParams struct {
 	items int
 	err   error
-}
-
-func testTelemetry(t *testing.T, id component.ID, testFunc func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool)) {
-	t.Run("WithOC", func(t *testing.T) {
-		tt, err := obsreporttest.SetupTelemetry(id)
-		require.NoError(t, err)
-		t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
-
-		testFunc(t, tt, false)
-	})
-
-	t.Run("WithOTel", func(t *testing.T) {
-		tt, err := obsreporttest.SetupTelemetry(id)
-		require.NoError(t, err)
-		t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
-
-		testFunc(t, tt, true)
-	})
 }
 
 func TestReceiveTraceDataOp(t *testing.T) {
@@ -65,7 +44,7 @@ func TestReceiveTraceDataOp(t *testing.T) {
 			{items: 42, err: nil},
 		}
 		for i, param := range params {
-			rec, err := newReceiver(ReceiverSettings{
+			rec, err := newReceiver(ObsReportSettings{
 				ReceiverID:             receiverID,
 				Transport:              transport,
 				ReceiverCreateSettings: tt.ToReceiverCreateSettings(),
@@ -112,7 +91,7 @@ func TestReceiveLogsOp(t *testing.T) {
 			{items: 42, err: nil},
 		}
 		for i, param := range params {
-			rec, err := newReceiver(ReceiverSettings{
+			rec, err := newReceiver(ObsReportSettings{
 				ReceiverID:             receiverID,
 				Transport:              transport,
 				ReceiverCreateSettings: tt.ToReceiverCreateSettings(),
@@ -160,7 +139,7 @@ func TestReceiveMetricsOp(t *testing.T) {
 			{items: 29, err: nil},
 		}
 		for i, param := range params {
-			rec, err := newReceiver(ReceiverSettings{
+			rec, err := newReceiver(ObsReportSettings{
 				ReceiverID:             receiverID,
 				Transport:              transport,
 				ReceiverCreateSettings: tt.ToReceiverCreateSettings(),
@@ -199,63 +178,6 @@ func TestReceiveMetricsOp(t *testing.T) {
 	})
 }
 
-func TestScrapeMetricsDataOp(t *testing.T) {
-	testTelemetry(t, receiverID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		parentCtx, parentSpan := tt.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
-		defer parentSpan.End()
-
-		params := []testParams{
-			{items: 23, err: partialErrFake},
-			{items: 29, err: errFake},
-			{items: 15, err: nil},
-		}
-		for i := range params {
-			scrp, err := newScraper(ScraperSettings{
-				ReceiverID:             receiverID,
-				Scraper:                scraperID,
-				ReceiverCreateSettings: tt.ToReceiverCreateSettings(),
-			}, useOtel)
-			require.NoError(t, err)
-			ctx := scrp.StartMetricsOp(parentCtx)
-			assert.NotNil(t, ctx)
-			scrp.EndMetricsOp(ctx, params[i].items, params[i].err)
-		}
-
-		spans := tt.SpanRecorder.Ended()
-		require.Equal(t, len(params), len(spans))
-
-		var scrapedMetricPoints, erroredMetricPoints int
-		for i, span := range spans {
-			assert.Equal(t, "scraper/"+receiverID.String()+"/"+scraperID.String()+"/MetricsScraped", span.Name())
-			switch {
-			case params[i].err == nil:
-				scrapedMetricPoints += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(0)})
-				assert.Equal(t, codes.Unset, span.Status().Code)
-			case errors.Is(params[i].err, errFake):
-				erroredMetricPoints += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(0)})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				assert.Equal(t, codes.Error, span.Status().Code)
-				assert.Equal(t, params[i].err.Error(), span.Status().Description)
-
-			case errors.Is(params[i].err, partialErrFake):
-				scrapedMetricPoints += params[i].items
-				erroredMetricPoints++
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ScrapedMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.ErroredMetricPointsKey, Value: attribute.Int64Value(1)})
-				assert.Equal(t, codes.Error, span.Status().Code)
-				assert.Equal(t, params[i].err.Error(), span.Status().Description)
-			default:
-				t.Fatalf("unexpected err param: %v", params[i].err)
-			}
-		}
-
-		require.NoError(t, obsreporttest.CheckScraperMetrics(tt, receiverID, scraperID, int64(scrapedMetricPoints), int64(erroredMetricPoints)))
-	})
-}
-
 func TestReceiveWithLongLivedCtx(t *testing.T) {
 	tt, err := obsreporttest.SetupTelemetry(receiverID)
 	require.NoError(t, err)
@@ -271,7 +193,7 @@ func TestReceiveWithLongLivedCtx(t *testing.T) {
 	for i := range params {
 		// Use a new context on each operation to simulate distinct operations
 		// under the same long lived context.
-		rec, rerr := NewReceiver(ReceiverSettings{
+		rec, rerr := NewObsReport(ObsReportSettings{
 			ReceiverID:             receiverID,
 			Transport:              transport,
 			LongLivedCtx:           true,
@@ -308,4 +230,22 @@ func TestReceiveWithLongLivedCtx(t *testing.T) {
 			t.Fatalf("unexpected error: %v", params[i].err)
 		}
 	}
+}
+
+func testTelemetry(t *testing.T, id component.ID, testFunc func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool)) {
+	t.Run("WithOC", func(t *testing.T) {
+		tt, err := obsreporttest.SetupTelemetry(id)
+		require.NoError(t, err)
+		t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
+
+		testFunc(t, tt, false)
+	})
+
+	t.Run("WithOTel", func(t *testing.T) {
+		tt, err := obsreporttest.SetupTelemetry(id)
+		require.NoError(t, err)
+		t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
+
+		testFunc(t, tt, true)
+	})
 }
